Updated logic to improve performance for user/team middleware

+++ server/models/team.js
@@ -24,15 +24,15 @@ const TeamSchema = new mongoose.Schema({
 });
 
 TeamSchema.pre('save', async function () {
-  if (this.isNew) {
+  if (this.isNew && this.status === 'new') {
     await Notification.create({
       notice: 'new',
       item: this._id,
       itemType: 'Team'
     });
   }
-  
-  if (this.isModified('status')) {
+
+  if (!this.isNew && this.isModified('status')) {
     if (this._status === 'new' && this.status !== this._status) {
       await Notification.findOneAndRemove({ item: this._id, notice: 'new' });
     }
